+++ base/utils.js
@@ -9,6 +9,6 @@ var hasModuleInstalled = function(name) {
 };
 
 module.exports = {
-    importOrderPluginInstalled: hasModuleInstalled("esling-plugin-import-order"),
+    importPluginInstalled: hasModuleInstalled("eslint-plugin-import"),
     reactPluginInstalled: hasModuleInstalled("eslint-plugin-react")
 };